Reset

+++ app/views/header.php
@@ -23,31 +23,31 @@
 
                 $arguments = explode('/', trim($_GET['url'], '/'));
 
-                if ($arguments[1]) {
+                if ($arguments[1] && $exist) {
                     echo '<li><a href="/agenda">Inicio</a></li>';
                 } else {
                     echo '<li>Inicio</li>';
                 }
 
-                if ($arguments[1] !== 'insert') {
+                if ($arguments[1] !== 'insert'  && $exist) {
                     echo '<li><a href="/agenda/insert">Insert</a></li>';
                 } else {
                     echo '<li>Insert</li>';
                 }
 
-                if ($arguments[1] !== 'delete') {
+                if ($arguments[1] !== 'delete' && $exist) {
                     echo '<li><a href="/agenda/delete">Delete</a></li>';
                 } else {
                     echo '<li>Delete</li>';
                 }
 
-                if ($arguments[1] !== 'search') {
+                if ($arguments[1] !== 'search' && $exist) {
                     echo '<li><a href="/agenda/search">Search</a></li>';
                 } else {
                     echo '<li>Search</li>';
                 }
 
-                if ($arguments[1] !== 'update') {
+                if ($arguments[1] !== 'update' && $exist) {
                     echo '<li><a href="/agenda/update">Update</a></li>';
                 } else {
                     echo '<li>Update</li>';
